admin integration test done!

+++ src/main/java/com/departement/fichedevoeux/repository/ModuleRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ModuleRepository extends JpaRepository <Module, Integer> {
+public interface ModuleRepository extends JpaRepository <Module, Long> {
 	
 	// Filtrer les modules en fonction du semestre, pallier et spécialité
     List<Module> findBySemestreAndPallierAndSpecialite(String semestre, String pallier, String specialite);
